Refactor some code
